refactor(fretboard): extract svg render layers
Split FretboardSVG.tsx JSX into five focused pure render layers:
FretboardDefs, FretboardShapeLayer, FretboardNoteLayer,
FretboardHitTargetLayer, and FretNumbersRow. FretboardSVG.tsx becomes
a thin orchestrator that handles hooks and prop distribution only.
